Squad based AI - Initial Commit - Move all default behaviours to seperate files
