+++ creep.Enforcer.js
@@ -1,15 +1,11 @@
-module.exports = function (creep) {
+module.exports = {
+    behaviours: {
+        default: require("creep.Enforcer.behaviour.Default")
+    },
 
-	var target = Game.getObjectById(Memory.Military.Target);
-	console.log(Memory.Military.Target);
-	console.log(target);
-    if(target.my === true) {
-        if(creep.pos.isNearTo(target)) {
-            creep.attack(target);
-        } else {
-            creep.moveTo(target);
-        }
-    } else {
-        creep.moveTo(target);
+    compute: function(creep) {
+        var desiredBehaviour = default;
+
+        return this.behaviours[desiredBehaviour](creep);
     }
-};
+}
